Fixed issue where logged in user could see edit and delete buttons for other users when viewing their profile.

+++ application/views/user_posts.php
@@ -20,8 +20,10 @@ foreach ($posts as $msg) :
 		<div class="my-well">
 			<h3><a href="/users/display/<?php echo $msg->user_id; ?>"><?php echo $msg->user_name; ?></a></h3>
 			<span>
+				<?php if ($this->session->userdata('user_id') == $msg->user_id) { ?>
                                 <a href="/posts/edit/<?php echo $msg->id; ?>" class="btn btn-mini btn-warning"><i class="icon-edit"></i></a>
                                 <a href="/posts/delete/<?php echo $msg->id; ?>" class="btn btn-mini btn-warning"><i class="icon-trash"></i></a>
+				<?php } ?>
                                 <button class="btn btn-mini btn-success"><i class="icon-thumbs-up"><?php if ($msg->karma_up > 0) { echo $msg->karma_up;} ?></i></a>
                                 <button class="btn btn-mini btn-danger"><i class="icon-thumbs-down"><?php if ($msg->karma_down > 0) { echo $msg->karma_down;} ?></i></a>
                         </span>
